[gh-#825] add delete connection call and update response code
