+++ packages/server/lib/controllers/config.controller.ts
@@ -308,7 +308,7 @@ class ConfigController {
 
             await configService.deleteProviderConfig(providerConfigKey, environmentId);
 
-            res.status(200).send();
+            res.status(204).send();
         } catch (err) {
             next(err);
         }
